ShopDetailBanner.tsx を責務別コンポーネントに分割する (#190)
2607行の巨大コンポーネントを5つのファイルに分割:
- ShopBannerHero.tsx: ヒーロー画像・ヘッダー・型定義
- PostCarousel.tsx: お知らせカルーセル
- CouponInfoCard.tsx: クーポン表示カード
- AiConsultPanel.tsx: AIチャットパネル
- KotodutePanel.tsx: ことづてパネル
