+++ app/(public)/map/components/ShopDetailBanner.tsx
@@ -1,27 +1,21 @@
-// app/(public)/map/components/ShopDetailBanner.tsx
 "use client";
 
 import { useState, useCallback, useEffect, useLayoutEffect, useMemo, useRef } from "react";
-import type { CSSProperties, ReactNode, RefObject } from "react";
+import type { CSSProperties, RefObject } from "react";
 import Image from "next/image";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import {
-  MessageSquarePlus,
   MapPin,
   ShoppingBag,
-  Clock,
   ChevronDown,
   ChevronLeft,
   ChevronRight,
   Instagram,
   Globe,
-  X as XIcon,
-  Pencil,
-  ArrowLeft,
-  ArrowRight,
-  Send,
+  XIcon,
   Sparkles,
+  MessageSquarePlus,
 } from "lucide-react";
 import { Shop } from "../data/shops";
 import { useAuth } from "../../../../lib/auth/AuthContext";
@@ -29,12 +23,16 @@ import { getShopBannerImage } from "../../../../lib/shopImages";
 import { useBag } from "../../../../lib/storage/BagContext";
 import { incrementBannerOpens } from "../../../../lib/storage/marketStats";
 import { ingredientCatalog, recipes } from "../../../../lib/recipes";
+import { loadKotodute, KOTODUTE_UPDATED_EVENT, type KotoduteNote } from "../../../../lib/kotoduteStorage";
 import {
-  KOTODUTE_UPDATED_EVENT,
-  loadKotodute,
-  saveKotodute,
-  type KotoduteNote,
-} from "../../../../lib/kotoduteStorage";
+  ShopBannerHero,
+  ShopBusinessInfoCard,
+  type ActivePostItem,
+} from "./ShopBannerHero";
+import { PostCarousel } from "./PostCarousel";
+import { CouponInfoCard } from "./CouponInfoCard";
+import { AiConsultPanel } from "./AiConsultPanel";
+import { KotodutePanel } from "./KotodutePanel";
 
 // ─── Theme presets ────────────────────────────────────────────────────────────
 const THEME_PRESETS = {
@@ -47,7 +45,6 @@ const THEME_PRESETS = {
 } as const;
 
 type ThemeKey = keyof typeof THEME_PRESETS;
-type BannerTheme = { bg: string; accent: string; text: string; border: string; light: string };
 type MainSurface = "summary" | "detail";
 type BannerSurface = MainSurface | "kotodute" | "ai";
 
@@ -82,12 +79,6 @@ type BagItem = {
   fromShopId?: number;
 };
 
-type ActivePostItem = {
-  text: string;
-  imageUrl?: string;
-  expiresAt: string;
-  createdAt?: string;
-};
 
 // ─── Constants ────────────────────────────────────────────────────────────────
 const STORAGE_KEY = "nicchyo-fridge-items";
@@ -103,73 +94,6 @@ const COLLAPSED_SUMMARY_OFFSET_PX = 10;
 const buildBagKey = (name: string, shopId?: number) =>
   `${name.trim().toLowerCase()}-${shopId ?? "any"}`;
 
-function formatBusinessHours(start?: string, end?: string) {
-  const trimmedStart = start?.trim();
-  const trimmedEnd = end?.trim();
-  if (trimmedStart && trimmedEnd) return `${trimmedStart} 〜 ${trimmedEnd}`;
-  if (trimmedStart) return `${trimmedStart}から`;
-  if (trimmedEnd) return `${trimmedEnd}まで`;
-  return null;
-}
-
-function ShopBusinessInfoCard({ shop, theme }: { shop: Shop; theme: BannerTheme }) {
-  const businessHours = formatBusinessHours(shop.businessHoursStart, shop.businessHoursEnd);
-  const scheduleText = shop.schedule?.trim() || "出店予定は未設定です";
-  const rainPolicyLabel =
-    shop.rainPolicy && shop.rainPolicy !== "undecided"
-      ? shop.rainPolicy === "outdoor"
-        ? "雨でも出店"
-        : shop.rainPolicy === "tent"
-          ? "雨でも出店（テント）"
-          : "雨天中止"
-      : null;
-
-  return (
-    <div
-      className="rounded-2xl border p-4 shadow-sm"
-      style={{ borderColor: theme.border, backgroundColor: theme.bg }}
-    >
-      <div className="flex items-center gap-3">
-        <div
-          className="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-white shadow-sm"
-          aria-hidden
-        >
-          <Clock className="h-4 w-4" style={{ color: theme.text }} />
-        </div>
-        <div className="min-w-0">
-          <p className="text-[11px] font-bold uppercase tracking-[0.18em]" style={{ color: theme.text }}>
-            営業情報
-          </p>
-          <p className="mt-0.5 text-sm font-semibold text-slate-700">
-            今日行く前に、ここだけ確認できます
-          </p>
-        </div>
-      </div>
-
-      <div className="mt-4 grid gap-3 sm:grid-cols-2">
-        <div className="rounded-2xl border border-white/80 bg-white px-3 py-3 shadow-sm">
-          <p className="text-[11px] font-bold uppercase tracking-[0.16em] text-slate-400">
-            出店予定
-          </p>
-          <p className="mt-1 text-sm font-semibold leading-relaxed text-slate-800">
-            {scheduleText}
-          </p>
-        </div>
-        <div className="rounded-2xl border border-white/80 bg-white px-3 py-3 shadow-sm">
-          <p className="text-[11px] font-bold uppercase tracking-[0.16em] text-slate-400">
-            営業時間
-          </p>
-          <p className="mt-1 text-sm font-semibold text-slate-800">
-            {businessHours || "営業時間は未設定です"}
-          </p>
-          {rainPolicyLabel && (
-            <p className="mt-1 text-xs font-medium text-slate-500">{rainPolicyLabel}</p>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
 
 function findIngredientMatch(name: string) {
   const lower = name.trim().toLowerCase();
@@ -221,397 +145,10 @@ function useCenterBounceTrigger(
   return isActive;
 }
 
-function formatRelativeTime(ts: number): string {
-  const diff = Date.now() - ts;
-  const mins = Math.floor(diff / 60000);
-  if (mins < 1) return "たった今";
-  if (mins < 60) return `${mins}分前`;
-  const hours = Math.floor(mins / 60);
-  if (hours < 24) return `${hours}時間前`;
-  const days = Math.floor(hours / 24);
-  return `${days}日前`;
-}
 
-// ─── Category fallback ────────────────────────────────────────────────────────
-const CATEGORY_FALLBACK: Record<string, { emoji: string; gradient: string }> = {
-  "食材":     { emoji: "🥦", gradient: "bg-gradient-to-br from-emerald-100 to-green-200" },
-  "加工食品": { emoji: "🍱", gradient: "bg-gradient-to-br from-amber-100 to-orange-200" },
-  "工芸品":   { emoji: "🎨", gradient: "bg-gradient-to-br from-purple-100 to-indigo-200" },
-  "植物":     { emoji: "🌿", gradient: "bg-gradient-to-br from-green-100 to-emerald-200" },
-  "飲食":     { emoji: "🍜", gradient: "bg-gradient-to-br from-orange-100 to-red-200" },
-};
 
-function ShopBannerHero({
-  shop,
-  bannerImage,
-  theme,
-  heroImageError,
-  onImageError,
-  mode,
-  isKotodute,
-  showProductPreview = false,
-  onEdit,
-  couponBadge,
-  couponStatus,
-  primaryCouponSetting,
-}: {
-  shop: Shop;
-  bannerImage: string;
-  theme: BannerTheme;
-  heroImageError: boolean;
-  onImageError: () => void;
-  mode: "compact" | "expanded";
-  isKotodute: boolean;
-  showProductPreview?: boolean;
-  onEdit?: () => void;
-  couponBadge?: React.ReactNode;
-  couponStatus?: "active" | "stamped" | "participating" | null;
-  primaryCouponSetting?: {
-    coupon_type_id: string;
-    coupon_type_name: string;
-    coupon_type_emoji: string;
-    coupon_type_amount: number;
-    min_purchase_amount: number;
-  } | null;
-}) {
-  if (mode === "compact") {
-    // クーポン状態に応じた右側カード
-    const couponCard =
-      couponStatus === "active" && primaryCouponSetting ? (
-        <div className="mr-10 flex h-[86px] w-[84px] shrink-0 flex-col items-center justify-center rounded-[18px] bg-green-500 text-white shadow-md">
-          <span className="text-[15px] leading-none">{primaryCouponSetting.coupon_type_emoji}</span>
-          <span className="mt-1 text-[22px] font-extrabold leading-none">
-            ¥{primaryCouponSetting.coupon_type_amount.toLocaleString()}
-          </span>
-          <span className="mt-0.5 text-[10px] font-bold opacity-90">引き</span>
-          <span className="mt-0.5 rounded-full bg-white/20 px-2 py-0.5 text-[9px] font-bold">
-            今すぐ使える
-          </span>
-        </div>
-      ) : couponStatus === "stamped" ? (
-        <div className="mr-10 flex h-[86px] w-[84px] shrink-0 flex-col items-center justify-center gap-0.5 rounded-[18px] border border-emerald-200 bg-emerald-50">
-          <span className="text-[22px]">✅</span>
-          <span className="text-center text-[10px] font-bold leading-tight text-emerald-700">
-            スタンプ
-            <br />
-            済み
-          </span>
-        </div>
-      ) : couponStatus === "participating" && primaryCouponSetting ? (
-        <div className="mr-10 flex h-[86px] w-[84px] shrink-0 flex-col items-center justify-center gap-0.5 rounded-[18px] border-[1.5px] border-dashed border-green-200 bg-green-50">
-          <span className="text-[18px]">🎟️</span>
-          <span className="text-center text-[10px] font-bold leading-tight text-green-700">
-            クーポン
-            <br />
-            使えます
-          </span>
-          <span className="mt-0.5 text-[9px] font-semibold text-green-500">
-            ¥{primaryCouponSetting.coupon_type_amount.toLocaleString()}引き
-          </span>
-        </div>
-      ) : null;
 
-    return (
-      <div className="flex items-center gap-3">
-        {/* サムネイル */}
-        <div className="relative h-16 w-16 shrink-0 overflow-hidden rounded-2xl border border-slate-200 bg-slate-100">
-          {heroImageError ? (
-            <div className="flex h-full w-full items-center justify-center bg-slate-100 text-xl">
-              {CATEGORY_FALLBACK[shop.category ?? ""]?.emoji ?? "🏪"}
-            </div>
-          ) : (
-            <Image
-              src={bannerImage}
-              alt={`${shop.name}の写真`}
-              fill
-              className="object-cover object-center"
-              onError={onImageError}
-            />
-          )}
-        </div>
 
-        {/* 店舗情報 */}
-        <div className={`min-w-0 flex-1 ${!couponCard ? "pr-12" : ""}`}>
-          <div className="flex flex-wrap items-center gap-1">
-            <span
-              className="rounded-full px-2 py-0.5 text-[10px] font-bold uppercase tracking-[0.12em]"
-              style={{ backgroundColor: theme.light, color: theme.text }}
-            >
-              {shop.category || "ショップ"}
-            </span>
-            <span className="rounded-full bg-emerald-50 px-2 py-0.5 text-[10px] font-bold text-emerald-700">
-              今日出店中
-            </span>
-            {!couponCard && couponBadge}
-          </div>
-          <h2 className="mt-1 line-clamp-2 text-[17px] font-extrabold leading-tight text-slate-900">
-            {shop.name}
-          </h2>
-          {!isKotodute && shop.catchphrase && (
-            <p className="mt-0.5 line-clamp-1 text-xs text-slate-500">{shop.catchphrase}</p>
-          )}
-          {/* クーポンがない場合のみ商品プレビュー */}
-          {!couponCard && showProductPreview && !isKotodute && shop.products.length > 0 && (
-            <div className="mt-1.5 flex gap-1.5 overflow-x-auto" style={{ scrollbarWidth: "none" }}>
-              {shop.products.slice(0, 4).map((product) => (
-                <span
-                  key={product}
-                  className="shrink-0 rounded-full border border-slate-200 bg-slate-50 px-2.5 py-0.5 text-[10px] font-medium text-slate-600"
-                >
-                  {product}
-                </span>
-              ))}
-              {shop.products.length > 4 && (
-                <span className="shrink-0 rounded-full bg-slate-100 px-2.5 py-0.5 text-[10px] font-medium text-slate-400">
-                  +{shop.products.length - 4}
-                </span>
-              )}
-            </div>
-          )}
-        </div>
-
-        {/* クーポンカード（右側） */}
-        {couponCard}
-      </div>
-    );
-  }
-
-  return (
-    <div className="relative h-56 overflow-hidden rounded-[28px] border border-slate-100 bg-slate-100 shadow-sm md:h-64 md:rounded-none md:border-0 md:shadow-none">
-      {heroImageError ? (
-        <div
-          className={`flex h-full w-full items-center justify-center ${
-            CATEGORY_FALLBACK[shop.category ?? ""]?.gradient ?? "bg-gradient-to-br from-slate-100 to-slate-200"
-          }`}
-        >
-          <span className="text-7xl">{CATEGORY_FALLBACK[shop.category ?? ""]?.emoji ?? "🏪"}</span>
-        </div>
-      ) : (
-        <Image
-          src={bannerImage}
-          alt={`${shop.name}の写真`}
-          fill
-          className="object-cover object-center"
-          priority
-          onError={onImageError}
-        />
-      )}
-      <div className="absolute inset-0 bg-gradient-to-t from-black/70 via-black/20 to-transparent" />
-      <div className="absolute bottom-0 left-0 right-0 px-5 pb-4">
-        <div className="flex items-end justify-between gap-3">
-          <div className="min-w-0 flex-1">
-            <div className="mb-2 flex flex-wrap items-center gap-2">
-              <span
-                className="rounded-full bg-white/18 px-2.5 py-1 text-[11px] font-bold uppercase tracking-[0.16em] text-white backdrop-blur-sm"
-              >
-                {shop.category || "ショップ"}
-              </span>
-              <span className="rounded-full bg-emerald-400/20 px-2.5 py-1 text-[11px] font-bold text-white backdrop-blur-sm">
-                今日出店中
-              </span>
-            </div>
-            <h2 className="text-3xl font-extrabold leading-tight text-white drop-shadow-md md:text-4xl">
-              {shop.name}
-            </h2>
-            {!isKotodute && shop.catchphrase && (
-              <p className="mt-1 text-sm font-medium text-white/80 drop-shadow">
-                {shop.catchphrase}
-              </p>
-            )}
-          </div>
-          {onEdit && !isKotodute && (
-            <button
-              type="button"
-              onClick={onEdit}
-              className="shrink-0 flex items-center gap-1.5 rounded-full bg-white/20 px-3 py-1.5 text-xs font-bold text-white backdrop-blur-sm transition hover:bg-white/30"
-            >
-              <Pencil className="h-3 w-3" />
-              編集
-            </button>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ShopSubviewHeader({
-  shop,
-  bannerImage,
-  theme,
-  heroImageError,
-  onImageError,
-  title,
-  titleIcon,
-  onBack,
-  onClose,
-  rightSlot,
-}: {
-  shop: Shop;
-  bannerImage: string;
-  theme: BannerTheme;
-  heroImageError: boolean;
-  onImageError: () => void;
-  title: string;
-  titleIcon?: ReactNode;
-  onBack: () => void;
-  onClose?: () => void;
-  rightSlot?: ReactNode;
-}) {
-  return (
-    <div
-      className="sticky top-0 z-10 shrink-0 border-b bg-white/95 px-3 py-3 backdrop-blur-sm"
-      style={{ borderColor: theme.border }}
-    >
-      <div className="flex items-center gap-2">
-        <button
-          type="button"
-          onClick={onBack}
-          className="pointer-events-auto flex items-center gap-1 rounded-full px-2.5 py-1.5 text-xs font-semibold text-slate-600 transition hover:bg-slate-100 active:scale-95"
-        >
-          <ArrowLeft className="h-3.5 w-3.5" />
-          戻る
-        </button>
-
-        <div className="flex flex-1 items-center justify-center gap-1.5 text-sm font-bold text-slate-900">
-          {titleIcon}
-          <span>{title}</span>
-        </div>
-
-        {rightSlot || onClose ? (
-          <div className="flex min-w-[3.5rem] items-center justify-end gap-1.5">
-            {rightSlot}
-            {onClose && (
-              <button
-                type="button"
-                onClick={onClose}
-                className="flex h-8 w-8 items-center justify-center rounded-full bg-slate-100 text-slate-600 shadow-sm transition hover:bg-slate-200"
-                aria-label="閉じる"
-              >
-                <XIcon className="h-4 w-4" />
-              </button>
-            )}
-          </div>
-        ) : (
-          <div className="w-14" />
-        )}
-      </div>
-
-      <div className="mt-3 rounded-2xl border border-slate-100 bg-slate-50 px-3 py-2.5">
-        <div className="flex items-center gap-3">
-          <div className="relative h-12 w-12 shrink-0 overflow-hidden rounded-2xl border border-slate-200 bg-slate-100">
-            {heroImageError ? (
-              <div className="flex h-full w-full items-center justify-center bg-slate-100 text-base">
-                {CATEGORY_FALLBACK[shop.category ?? ""]?.emoji ?? "🏪"}
-              </div>
-            ) : (
-              <Image
-                src={bannerImage}
-                alt={`${shop.name}の写真`}
-                fill
-                className="object-cover object-center"
-                onError={onImageError}
-              />
-            )}
-          </div>
-          <div className="min-w-0 flex-1">
-            <div className="flex items-center gap-1.5">
-              <span
-                className="rounded-full px-2 py-0.5 text-[10px] font-bold uppercase tracking-[0.12em]"
-                style={{ backgroundColor: theme.light, color: theme.text }}
-              >
-                {shop.category || "ショップ"}
-              </span>
-              <span className="rounded-full bg-emerald-50 px-2 py-0.5 text-[10px] font-bold text-emerald-700">
-                今日出店中
-              </span>
-            </div>
-            <p className="mt-1 line-clamp-1 text-sm font-bold text-slate-900">{shop.name}</p>
-            {shop.catchphrase && (
-              <p className="line-clamp-1 text-xs text-slate-500">{shop.catchphrase}</p>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function BannerActivePostsCard({
-  activePosts,
-  theme,
-  currentPostIndex,
-  isActivePostCentered,
-  activePostRef,
-  activePostCarouselRef,
-}: {
-  activePosts: ActivePostItem[];
-  theme: BannerTheme;
-  currentPostIndex: number;
-  isActivePostCentered: boolean;
-  activePostRef: RefObject<HTMLDivElement>;
-  activePostCarouselRef: RefObject<HTMLDivElement>;
-}) {
-  return (
-    <div
-      ref={activePostRef}
-      className={`overflow-hidden rounded-2xl border shadow-sm ${isActivePostCentered ? "center-bounce-in" : ""}`}
-      style={{ borderColor: theme.border }}
-    >
-      <div className="flex items-center gap-2 px-4 py-2.5" style={{ backgroundColor: theme.light }}>
-        <span className="text-base">📢</span>
-        <span className="text-sm font-bold" style={{ color: theme.text }}>今日のお知らせ</span>
-        {activePosts.length > 1 && (
-          <div className="ml-auto flex gap-1">
-            {activePosts.map((_, i) => (
-              <div
-                key={i}
-                className="h-1.5 w-1.5 rounded-full transition-colors"
-                style={{ backgroundColor: i === currentPostIndex ? theme.accent : theme.border }}
-              />
-            ))}
-          </div>
-        )}
-      </div>
-
-      <div ref={activePostCarouselRef} className="flex snap-x snap-mandatory overflow-x-hidden scroll-smooth">
-        {activePosts.map((post, index) => (
-          <article key={index} className="w-full shrink-0 snap-center">
-            {post.imageUrl && (
-              // eslint-disable-next-line @next/next/no-img-element
-              <img src={post.imageUrl} alt="お知らせ画像" className="h-48 w-full object-cover" />
-            )}
-            <div className="px-4 py-3">
-              <p className="whitespace-pre-wrap text-base leading-relaxed text-slate-800">{post.text}</p>
-              <div className="mt-2 flex items-center justify-between text-xs text-slate-400">
-                <span>
-                  {(() => {
-                    const diff = new Date(post.expiresAt).getTime() - Date.now();
-                    if (diff <= 0) return "期限切れ";
-                    const h = Math.floor(diff / 3600000);
-                    const m = Math.floor((diff % 3600000) / 60000);
-                    return h > 0 ? `あと${h}時間` : `あと${m}分`;
-                  })()}
-                </span>
-                {post.createdAt && (
-                  <span>
-                    {new Intl.DateTimeFormat("ja-JP", {
-                      timeZone: "Asia/Tokyo",
-                      month: "numeric",
-                      day: "numeric",
-                      hour: "numeric",
-                      minute: "2-digit",
-                    }).format(new Date(post.createdAt))}
-                  </span>
-                )}
-              </div>
-            </div>
-          </article>
-        ))}
-      </div>
-    </div>
-  );
-}
 
 // ─── Main Component ───────────────────────────────────────────────────────────
 export default function ShopDetailBanner({
@@ -1379,7 +916,7 @@ export default function ShopDetailBanner({
             <div className="rounded-[30px] border border-slate-200 bg-white px-5 py-5 shadow-sm">
               <div className="space-y-5">
                 {activePosts.length > 0 && (
-                  <BannerActivePostsCard
+                  <PostCarousel
                     activePosts={activePosts}
                     theme={theme}
                     currentPostIndex={currentPostIndex}
@@ -1663,7 +1200,7 @@ export default function ShopDetailBanner({
               TODAY'S ANNOUNCEMENT — Rich card
           ════════════════════════════════════════════════════════════════ */}
           {!isKotodute && activePosts.length > 0 && !isMobileOverlay && (
-            <BannerActivePostsCard
+            <PostCarousel
               activePosts={activePosts}
               theme={theme}
               currentPostIndex={currentPostIndex}
@@ -1851,7 +1388,6 @@ export default function ShopDetailBanner({
             theme={theme}
             onBack={handleBackToMain}
             onClose={isMobileOverlay ? onClose : undefined}
-            isActive={surface === "ai"}
           />
         </div>
           </div>
@@ -1874,734 +1410,5 @@ export default function ShopDetailBanner({
   );
 }
 
-// ─── Coupon Info Card ─────────────────────────────────────────────────────────
-type CouponSetting = {
-  coupon_type_id: string;
-  coupon_type_name: string;
-  coupon_type_emoji: string;
-  coupon_type_amount: number;
-  min_purchase_amount: number;
-};
 
-function CouponInfoCard({
-  setting,
-  allSettings,
-  couponStatus,
-  activeCouponTypeId,
-}: {
-  setting: CouponSetting;
-  allSettings: CouponSetting[];
-  couponStatus: "active" | "stamped" | "participating";
-  activeCouponTypeId?: string;
-}) {
-  if (couponStatus === "active") {
-    return (
-      <div className="rounded-xl border-2 border-green-300 bg-green-50 p-4">
-        <div className="flex items-center gap-2 mb-2">
-          <span className="text-base">{setting.coupon_type_emoji}</span>
-          <span className="text-sm font-bold text-green-800">{setting.coupon_type_name}</span>
-          <span className="rounded-full bg-green-600 px-2 py-0.5 text-[10px] font-bold text-white">
-            今すぐ使える
-          </span>
-        </div>
-        <p className="text-2xl font-extrabold text-green-600 mb-0.5">
-          {setting.coupon_type_amount.toLocaleString()}円引き
-        </p>
-        {setting.min_purchase_amount > 0 && (
-          <p className="text-xs text-slate-500">
-            {setting.min_purchase_amount.toLocaleString()}円以上のご購入で適用
-          </p>
-        )}
-        {allSettings.length > 1 && (
-          <div className="mt-2 flex flex-wrap gap-1.5">
-            {allSettings.filter((s) => s.coupon_type_id !== activeCouponTypeId).map((s) => (
-              <span
-                key={s.coupon_type_id}
-                className="rounded-full border border-green-200 bg-white px-2 py-0.5 text-[11px] text-green-700"
-              >
-                {s.coupon_type_emoji} {s.coupon_type_name}も対応
-              </span>
-            ))}
-          </div>
-        )}
-        <Link
-          href="/coupons"
-          className="mt-3 flex w-full items-center justify-center gap-1.5 rounded-xl bg-green-500 py-2.5 text-sm font-bold text-white transition hover:bg-green-600"
-        >
-          クーポンを確認する
-          <ArrowRight className="h-3.5 w-3.5" />
-        </Link>
-      </div>
-    );
-  }
 
-  if (couponStatus === "stamped") {
-    return (
-      <div className="rounded-xl border border-emerald-200 bg-emerald-50 p-4">
-        <div className="flex items-center gap-2">
-          <span className="text-base">{setting.coupon_type_emoji}</span>
-          <span className="text-sm font-semibold text-emerald-800">{setting.coupon_type_name}</span>
-          <span className="rounded-full bg-emerald-600 px-2 py-0.5 text-[10px] font-bold text-white">
-            ✅ 本日スタンプ済み
-          </span>
-        </div>
-        <p className="mt-1.5 text-xs text-emerald-700">
-          本日はこのお店のスタンプを取得済みです
-        </p>
-      </div>
-    );
-  }
-
-  // "participating" — クーポンは未保有
-  return (
-    <div className="rounded-xl border border-dashed border-green-200 bg-green-50/50 p-4">
-      <div className="flex items-center gap-2 mb-1.5">
-        <span className="text-base">{setting.coupon_type_emoji}</span>
-        <span className="text-sm font-semibold text-green-800">
-          {setting.coupon_type_name}クーポン対応
-        </span>
-      </div>
-      <p className="text-xs text-slate-500 mb-3">
-        クーポンページでQRコードを確認できます
-      </p>
-      <Link
-        href="/coupons"
-        className="flex w-full items-center justify-center gap-1.5 rounded-xl border border-green-300 bg-white py-2.5 text-sm font-semibold text-green-700 transition hover:bg-green-50"
-      >
-        クーポンを確認する
-        <ArrowRight className="h-3.5 w-3.5" />
-      </Link>
-    </div>
-  );
-}
-
-// ─── AI Consult Panel ────────────────────────────────────────────────────────
-type ChatMsg = { role: "user" | "assistant"; text: string };
-
-const AI_SUGGESTED_PROMPTS = [
-  { icon: "🛍️", text: "どんな商品がありますか？" },
-  { icon: "⭐", text: "おすすめはなんですか？" },
-  { icon: "🕐", text: "何時まで営業していますか？" },
-  { icon: "💴", text: "支払いはカードで払えますか？" },
-  { icon: "🌧️", text: "雨の日でも出店していますか？" },
-];
-
-function AiConsultPanel({
-  shop,
-  bannerImage,
-  heroImageError,
-  theme,
-  onBack,
-  onClose,
-  isActive: _isActive,
-}: {
-  shop: Shop;
-  bannerImage: string;
-  heroImageError: boolean;
-  theme: BannerTheme;
-  onBack: () => void;
-  onClose?: () => void;
-  isActive: boolean;
-}) {
-  const [messages, setMessages] = useState<ChatMsg[]>([]);
-  const [input, setInput] = useState("");
-  const [streaming, setStreaming] = useState(false);
-  const messagesEndRef = useRef<HTMLDivElement | null>(null);
-  const textareaRef = useRef<HTMLTextAreaElement | null>(null);
-  const abortRef = useRef<AbortController | null>(null);
-
-  // ショップが変わったらリセット
-  useEffect(() => {
-    setMessages([]);
-    setInput("");
-    setStreaming(false);
-    abortRef.current?.abort();
-  }, [shop.id]);
-
-  // アンマウント時にストリーミングを中断
-  useEffect(() => {
-    return () => { abortRef.current?.abort(); };
-  }, []);
-
-  // 新メッセージが追加されたら最下部へスクロール
-  useEffect(() => {
-    messagesEndRef.current?.scrollIntoView({ behavior: "smooth" });
-  }, [messages]);
-
-  // textarea の高さを内容に合わせて自動調整
-  const adjustTextarea = useCallback((el: HTMLTextAreaElement) => {
-    el.style.height = "auto";
-    el.style.height = `${Math.min(el.scrollHeight, 120)}px`;
-  }, []);
-
-  const handleInputChange = useCallback(
-    (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-      setInput(e.target.value);
-      adjustTextarea(e.target);
-    },
-    [adjustTextarea]
-  );
-
-  const sendMessage = useCallback(
-    async (text: string) => {
-      const trimmed = text.trim();
-      if (!trimmed || streaming) return;
-
-      const history = messages;
-      setMessages((prev) => [
-        ...prev,
-        { role: "user", text: trimmed },
-        { role: "assistant", text: "" },
-      ]);
-      setInput("");
-      if (textareaRef.current) {
-        textareaRef.current.style.height = "auto";
-      }
-      setStreaming(true);
-
-      const ctrl = new AbortController();
-      abortRef.current = ctrl;
-
-      try {
-        const res = await fetch("/api/grandma/shop-chat", {
-          method: "POST",
-          headers: { "Content-Type": "application/json" },
-          signal: ctrl.signal,
-          body: JSON.stringify({
-            shopName: shop.name,
-            shopContext: {
-              category: shop.category,
-              catchphrase: shop.catchphrase,
-              shopStrength: shop.shopStrength,
-              products: shop.products,
-              chome: shop.chome,
-            },
-            history,
-            text: trimmed,
-          }),
-        });
-        if (!res.ok || !res.body) throw new Error("upstream error");
-
-        const reader = res.body.getReader();
-        const decoder = new TextDecoder();
-        while (true) {
-          const { done, value } = await reader.read();
-          if (done) break;
-          const chunk = decoder.decode(value, { stream: true });
-          setMessages((prev) => {
-            const copy = [...prev];
-            const last = copy[copy.length - 1];
-            if (last?.role === "assistant") {
-              copy[copy.length - 1] = { ...last, text: last.text + chunk };
-            }
-            return copy;
-          });
-        }
-      } catch (err: unknown) {
-        if (err instanceof Error && err.name !== "AbortError") {
-          setMessages((prev) => {
-            const copy = [...prev];
-            const last = copy[copy.length - 1];
-            if (last?.role === "assistant" && !last.text) {
-              copy[copy.length - 1] = {
-                ...last,
-                text: "ごめんよ、うまく答えられんかったわ…もう一回試してみてね。",
-              };
-            }
-            return copy;
-          });
-        }
-      } finally {
-        setStreaming(false);
-        abortRef.current = null;
-      }
-    },
-    [messages, shop, streaming]
-  );
-
-  const handleAbort = useCallback(() => {
-    abortRef.current?.abort();
-    setStreaming(false);
-  }, []);
-
-  const handleSubmit = useCallback(
-    () => sendMessage(input),
-    [input, sendMessage]
-  );
-
-  const handleKeyDown = useCallback(
-    (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
-      // IME確定中は無視（日本語入力対応）
-      if (e.nativeEvent.isComposing) return;
-      if (e.key === "Enter" && !e.shiftKey) {
-        e.preventDefault();
-        handleSubmit();
-      }
-    },
-    [handleSubmit]
-  );
-
-  const isEmpty = messages.length === 0;
-  const lastMsg = messages[messages.length - 1];
-  const isTyping =
-    streaming && lastMsg?.role === "assistant" && !lastMsg.text;
-
-  return (
-    <div className="flex h-full flex-col overflow-hidden">
-      <ShopSubviewHeader
-        shop={shop}
-        bannerImage={bannerImage}
-        heroImageError={heroImageError}
-        onImageError={() => {}}
-        theme={theme}
-        title="AIに相談する"
-        titleIcon={<Sparkles className="h-3.5 w-3.5" style={{ color: theme.accent }} />}
-        onBack={onBack}
-        onClose={onClose}
-        rightSlot={
-          !isEmpty ? (
-            <button
-              type="button"
-              onClick={() => { setMessages([]); setInput(""); }}
-              className="text-xs font-semibold text-slate-400 transition hover:text-slate-600 px-2 py-1.5 rounded-full hover:bg-slate-100"
-            >
-              クリア
-            </button>
-          ) : undefined
-        }
-      />
-
-      {/* ── メッセージエリア ───────────────────────────────────────── */}
-      <div className="flex-1 overflow-y-auto overscroll-contain">
-        {isEmpty ? (
-          /* ── ウェルカム画面 ── */
-          <div className="flex flex-col items-center px-5 py-6 gap-5">
-            {/* キャラクター */}
-            <div className="flex flex-col items-center gap-3">
-              <div
-                className="relative flex h-20 w-20 items-center justify-center rounded-full shadow-md"
-                style={{ backgroundColor: theme.light }}
-              >
-                <Image
-                  src="/images/obaasan_transparent.png"
-                  alt="にちよさん"
-                  width={56}
-                  height={56}
-                  className="h-14 w-14 opacity-85"
-                />
-                {/* ステータスバッジ */}
-                <span
-                  className="absolute bottom-0.5 right-0.5 flex h-5 w-5 items-center justify-center rounded-full text-[9px] font-bold text-white shadow"
-                  style={{ backgroundColor: theme.accent }}
-                >
-                  AI
-                </span>
-              </div>
-              <div className="text-center">
-                <p className="text-[15px] font-bold text-slate-800 leading-snug">
-                  {shop.name}のことなら何でも！
-                </p>
-                <p className="mt-1 text-xs text-slate-400">
-                  土佐弁で親切にお答えするがよ〜
-                </p>
-              </div>
-            </div>
-
-            {/* 区切り */}
-            <div className="flex w-full items-center gap-3">
-              <div className="flex-1 border-t border-slate-100" />
-              <p className="text-[11px] font-semibold text-slate-300 uppercase tracking-wider">
-                よく聞かれる質問
-              </p>
-              <div className="flex-1 border-t border-slate-100" />
-            </div>
-
-            {/* サジェストプロンプト */}
-            <div className="w-full space-y-2">
-              {AI_SUGGESTED_PROMPTS.map(({ icon, text }) => (
-                <button
-                  key={text}
-                  type="button"
-                  onClick={() => sendMessage(text)}
-                  className="flex w-full items-center gap-3 rounded-2xl border px-4 py-3 text-sm font-medium transition hover:opacity-80 active:scale-[0.98] text-left"
-                  style={{
-                    borderColor: theme.border,
-                    backgroundColor: theme.bg,
-                    color: theme.text,
-                  }}
-                >
-                  <span className="text-base leading-none">{icon}</span>
-                  <span className="flex-1">{text}</span>
-                  <ChevronRight
-                    className="h-3.5 w-3.5 shrink-0 opacity-40"
-                    style={{ color: theme.text }}
-                  />
-                </button>
-              ))}
-            </div>
-
-            <p className="text-[10px] text-slate-300 text-center leading-relaxed">
-              AIの回答はお店情報に基づく参考情報です。<br />
-              内容の正確性を保証するものではありません。
-            </p>
-          </div>
-        ) : (
-          /* ── チャット表示 ── */
-          <div className="space-y-3 px-4 py-5 pb-3">
-            {messages.map((msg, i) => {
-              const isUser = msg.role === "user";
-              const isLastAssistant =
-                !isUser && i === messages.length - 1;
-              const showTyping = isTyping && isLastAssistant;
-
-              return (
-                <div
-                  key={i}
-                  className={`flex items-end gap-2.5 animate-in fade-in slide-in-from-bottom-2 duration-200 ${
-                    isUser ? "flex-row-reverse" : "flex-row"
-                  }`}
-                >
-                  {/* アシスタントアバター */}
-                  {!isUser && (
-                    <div
-                      className="mb-0.5 flex h-8 w-8 shrink-0 items-center justify-center rounded-full shadow-sm"
-                      style={{ backgroundColor: theme.light }}
-                    >
-                      <Image
-                        src="/images/obaasan_transparent.png"
-                        alt="にちよさん"
-                        width={24}
-                        height={24}
-                        className="h-6 w-6 opacity-85"
-                      />
-                    </div>
-                  )}
-
-                  {/* バブル */}
-                  <div
-                    className={`max-w-[78%] rounded-2xl px-4 py-3 text-sm leading-relaxed shadow-sm ${
-                      isUser
-                        ? "rounded-br-sm bg-slate-900 text-white"
-                        : "rounded-bl-sm border text-slate-800"
-                    }`}
-                    style={
-                      !isUser
-                        ? {
-                            borderColor: theme.border,
-                            backgroundColor: theme.bg,
-                          }
-                        : undefined
-                    }
-                  >
-                    {showTyping ? (
-                      /* タイピングインジケーター */
-                      <span className="inline-flex items-center gap-1 py-0.5 px-0.5">
-                        {[0, 160, 320].map((delay) => (
-                          <span
-                            key={delay}
-                            className="h-2 w-2 rounded-full animate-bounce"
-                            style={{
-                              backgroundColor: theme.accent,
-                              animationDelay: `${delay}ms`,
-                              opacity: 0.7,
-                            }}
-                          />
-                        ))}
-                      </span>
-                    ) : (
-                      <span className="whitespace-pre-wrap">{msg.text}</span>
-                    )}
-                  </div>
-                </div>
-              );
-            })}
-            {/* スクロール追従用アンカー */}
-            <div ref={messagesEndRef} />
-          </div>
-        )}
-      </div>
-
-      {/* ── 入力エリア ────────────────────────────────────────────── */}
-      <div
-        className="shrink-0 border-t bg-white px-3 pt-2 pb-3"
-        style={{ borderColor: theme.border }}
-      >
-        {/* 生成停止ボタン（ストリーミング中のみ） */}
-        {streaming && (
-          <div className="mb-2 flex justify-center">
-            <button
-              type="button"
-              onClick={handleAbort}
-              className="flex items-center gap-1.5 rounded-full border border-slate-200 bg-white px-3.5 py-1.5 text-xs font-semibold text-slate-500 shadow-sm transition hover:bg-slate-50 active:scale-95"
-            >
-              <span
-                className="h-2.5 w-2.5 rounded-sm"
-                style={{ backgroundColor: theme.accent }}
-              />
-              生成を停止
-            </button>
-          </div>
-        )}
-
-        {/* テキストエリア + 送信ボタン */}
-        <div
-          className="flex items-end gap-2 rounded-2xl border bg-slate-50 px-3.5 py-2.5 transition-shadow focus-within:shadow-md focus-within:bg-white"
-          style={{ borderColor: theme.border }}
-        >
-          <textarea
-            ref={textareaRef}
-            value={input}
-            onChange={handleInputChange}
-            onKeyDown={handleKeyDown}
-            placeholder={
-              streaming ? "回答中…" : "質問を入力（Shift+Enterで改行）"
-            }
-            disabled={streaming}
-            rows={1}
-            className="flex-1 resize-none bg-transparent text-sm text-slate-800 placeholder-slate-400 focus:outline-none disabled:cursor-not-allowed disabled:opacity-40"
-            style={{
-              lineHeight: "1.5",
-              maxHeight: 120,
-              overflowY: "auto",
-            }}
-          />
-          <button
-            type="button"
-            onClick={handleSubmit}
-            disabled={!input.trim() || streaming}
-            className="mb-0.5 flex h-9 w-9 shrink-0 items-center justify-center rounded-xl text-white shadow transition disabled:opacity-30 hover:opacity-90 active:scale-95"
-            style={{ backgroundColor: theme.accent }}
-            aria-label="送信"
-          >
-            <Send className="h-3.5 w-3.5" />
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ─── Kotodute characters ──────────────────────────────────────────────────────
-const KOTODUTE_CHARACTERS = [
-  { id: "sakura", emoji: "🌸", name: "さくら" },
-  { id: "kon",    emoji: "🦊", name: "こん" },
-  { id: "nami",   emoji: "🌊", name: "なみ" },
-  { id: "shiro",  emoji: "🏯", name: "しろ" },
-] as const;
-type KotoduteCharacterId = typeof KOTODUTE_CHARACTERS[number]["id"];
-const CHAR_STORAGE_KEY = "nicchyo-kotodute-character";
-
-// ─── Kotodute Panel (2nd slide) ───────────────────────────────────────────────
-function KotodutePanel({
-  shop,
-  bannerImage,
-  heroImageError,
-  theme,
-  onBack,
-  onClose,
-}: {
-  shop: Shop;
-  bannerImage: string;
-  heroImageError: boolean;
-  theme: BannerTheme;
-  onBack: () => void;
-  onClose?: () => void;
-}) {
-  const [allNotes, setAllNotes] = useState<KotoduteNote[]>([]);
-  const [text, setText] = useState("");
-  const [submitted, setSubmitted] = useState(false);
-  const [selectedCharId, setSelectedCharId] = useState<KotoduteCharacterId>(() => {
-    if (typeof window === "undefined") return "sakura";
-    return (localStorage.getItem(CHAR_STORAGE_KEY) as KotoduteCharacterId) ?? "sakura";
-  });
-  const textareaRef = useRef<HTMLTextAreaElement | null>(null);
-
-  const selectedChar = KOTODUTE_CHARACTERS.find((c) => c.id === selectedCharId) ?? KOTODUTE_CHARACTERS[0];
-
-  const notes = useMemo(
-    () =>
-      allNotes
-        .filter((n) => n.shopId === shop.id)
-        .sort((a, b) => b.createdAt - a.createdAt),
-    [allNotes, shop.id]
-  );
-
-  useEffect(() => {
-    setAllNotes(loadKotodute());
-    setText("");
-    setSubmitted(false);
-  }, [shop.id]);
-
-  useEffect(() => {
-    if (typeof window === "undefined") return;
-    const handler = () => setAllNotes(loadKotodute());
-    const storageHandler = (e: StorageEvent) => {
-      if (e.key === "nicchyo-kotodute-notes") handler();
-    };
-    window.addEventListener(KOTODUTE_UPDATED_EVENT, handler);
-    window.addEventListener("storage", storageHandler);
-    return () => {
-      window.removeEventListener(KOTODUTE_UPDATED_EVENT, handler);
-      window.removeEventListener("storage", storageHandler);
-    };
-  }, []);
-
-  const handleCharSelect = useCallback((id: KotoduteCharacterId) => {
-    setSelectedCharId(id);
-    localStorage.setItem(CHAR_STORAGE_KEY, id);
-  }, []);
-
-  const handleSubmit = useCallback(() => {
-    const body = text.trim();
-    if (!body) return;
-    const next: KotoduteNote = {
-      id: crypto.randomUUID(),
-      shopId: shop.id,
-      text: body,
-      createdAt: Date.now(),
-      authorEmoji: selectedChar.emoji,
-    };
-    const updated = [next, ...loadKotodute()];
-    saveKotodute(updated);
-    setAllNotes(updated);
-    setText("");
-    setSubmitted(true);
-    setTimeout(() => setSubmitted(false), 2000);
-  }, [text, shop.id, selectedChar]);
-
-  const handleKeyDown = useCallback(
-    (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
-      if ((e.metaKey || e.ctrlKey) && e.key === "Enter") handleSubmit();
-    },
-    [handleSubmit]
-  );
-
-  return (
-    <div className="flex h-full flex-col">
-      <ShopSubviewHeader
-        shop={shop}
-        bannerImage={bannerImage}
-        heroImageError={heroImageError}
-        onImageError={() => {}}
-        theme={theme}
-        title="ことづて"
-        onBack={onBack}
-        onClose={onClose}
-        rightSlot={
-          notes.length > 0 ? (
-            <span
-              className="rounded-full px-2 py-0.5 text-[11px] font-bold"
-              style={{ backgroundColor: theme.light, color: theme.text }}
-            >
-              {notes.length}
-            </span>
-          ) : undefined
-        }
-      />
-
-      {/* ── Compose area ────────────────────────────────────────────────────── */}
-      <div className="border-b px-4 py-4" style={{ borderColor: theme.border }}>
-        {/* Character selector */}
-        <div className="mb-3">
-          <p className="mb-1.5 text-[11px] font-semibold uppercase tracking-wider" style={{ color: theme.text }}>
-            投稿キャラを選ぶ
-          </p>
-          <div className="flex gap-2">
-            {KOTODUTE_CHARACTERS.map((char) => (
-              <button
-                key={char.id}
-                type="button"
-                onClick={() => handleCharSelect(char.id)}
-                className={`flex flex-1 flex-col items-center gap-0.5 rounded-xl py-2 text-xs font-bold transition active:scale-95 ${
-                  selectedCharId === char.id
-                    ? "shadow-sm ring-2"
-                    : "opacity-50 hover:opacity-80"
-                }`}
-                style={{
-                  backgroundColor: selectedCharId === char.id ? theme.light : "transparent",
-                  border: `1px solid ${theme.border}`,
-                  ...(selectedCharId === char.id ? { outline: `2px solid ${theme.accent}`, outlineOffset: "1px" } : {}),
-                }}
-              >
-                <span className="text-xl leading-none">{char.emoji}</span>
-                <span style={{ color: theme.text }}>{char.name}</span>
-              </button>
-            ))}
-          </div>
-        </div>
-
-        <p className="mb-2 text-xs font-semibold" style={{ color: theme.text }}>
-          {shop.name}へひとことメモ
-        </p>
-        <textarea
-          ref={textareaRef}
-          value={text}
-          onChange={(e) => setText(e.target.value)}
-          onKeyDown={handleKeyDown}
-          placeholder="おすすめや感想をひとこと…"
-          rows={3}
-          className="w-full resize-none rounded-xl border bg-slate-50 px-3 py-2.5 text-sm text-slate-800 placeholder-slate-400 transition focus:bg-white focus:outline-none focus:ring-2"
-          style={{
-            borderColor: theme.border,
-            // @ts-expect-error — CSSProperties doesn't include Tailwind custom vars
-            "--tw-ring-color": theme.accent + "55",
-          }}
-        />
-        <button
-          type="button"
-          onClick={handleSubmit}
-          disabled={!text.trim()}
-          className="mt-2.5 flex w-full items-center justify-center gap-2 rounded-xl py-2.5 text-sm font-bold text-white shadow-sm transition disabled:opacity-40 hover:opacity-90 active:scale-[0.98]"
-          style={{ backgroundColor: submitted ? "#22c55e" : theme.accent }}
-        >
-          {submitted ? (
-            "✓ 投稿しました！"
-          ) : (
-            <>
-              <Send className="h-3.5 w-3.5" />
-              投稿する
-            </>
-          )}
-        </button>
-      </div>
-
-      {/* ── Notes list ──────────────────────────────────────────────────────── */}
-      {notes.length === 0 ? (
-        <div className="flex flex-1 flex-col items-center justify-center gap-3 px-6 py-12 text-center">
-          <div
-            className="flex h-14 w-14 items-center justify-center rounded-full"
-            style={{ backgroundColor: theme.light }}
-          >
-            <MessageSquarePlus className="h-6 w-6" style={{ color: theme.accent }} />
-          </div>
-          <p className="text-sm font-bold text-slate-700">まだコメントがありません</p>
-          <p className="text-xs text-slate-400">最初の一言を投稿してみましょう！</p>
-        </div>
-      ) : (
-        <div className="space-y-3 px-4 py-4">
-          {notes.map((note) => (
-            <div key={note.id} className="flex items-start gap-2.5">
-              <div
-                className="mt-0.5 flex h-9 w-9 flex-shrink-0 items-center justify-center rounded-full text-xl shadow-sm"
-                style={{ backgroundColor: theme.light, border: `1.5px solid ${theme.border}` }}
-              >
-                {note.authorEmoji ?? "💬"}
-              </div>
-              <div
-                className="flex-1 rounded-2xl border px-3 py-2.5 text-sm"
-                style={{ borderColor: theme.border, backgroundColor: theme.bg }}
-              >
-              <p className="leading-relaxed text-slate-800">
-                {note.text.replace(KOTODUTE_TAG_REGEX, "").trim()}
-              </p>
-              <p className="mt-1.5 text-[11px]" style={{ color: theme.text, opacity: 0.6 }}>
-                {formatRelativeTime(note.createdAt)}
-              </p>
-              </div>
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
